+++ src/application/subito.rs
@@ -31,6 +31,20 @@ impl<Q, S, N> Subito<Q, S, N> {
     }
 }
 
+impl<Q, S, N> Subito<Q, S, N>
+where
+    N: NotificationApi,
+    Q: QueryApi,
+{
+    async fn notify(&self, result: &ItemResult) {
+        let _ = self.notification_api.notify(result).await;
+    }
+
+    async fn add_items(&self, items: Vec<ItemResult>) {
+        let _ = self.query_api.lock().await.add_items(items.clone()).await;
+    }
+}
+
 #[async_trait]
 impl<Q, S, N> ApplicationApi for Subito<Q, S, N>
 where
@@ -68,27 +82,33 @@ where
 
         let items = self.query_api.lock().await.fetch_all_items().await?;
         let mut results_to_write: Vec<ItemResult> = vec![];
+        let mut notification_handlers = vec![];
 
         for result in &results {
             if !items.contains(&result.get_uri()) {
                 results_to_write.push((*Arc::clone(result)).clone());
                 if notify.unwrap_or(true) {
-                    self.notification_api.notify(format!("{result}")).await?;
+                    notification_handlers.push(self.notify(result));
                 }
             }
         }
 
-        self.query_api
-            .lock()
-            .await
-            .add_items(results_to_write)
-            .await?;
+        tokio::join!(
+            futures::future::join_all(notification_handlers),
+            self.add_items(results_to_write)
+        );
 
         Ok(results)
     }
 
-    async fn add_user(&self, id: String) -> Result<(), Box<dyn Error>> {
+    async fn add_user(&self, id: &str) -> Result<(), Box<dyn Error>> {
         self.notification_api.add_user(id).await?;
+        Ok(())
+    }
+
+    async fn reset(&self) -> Result<(), Box<dyn Error>> {
+        self.query_api.lock().await.reset().await?;
+        self.notification_api.reset().await?;
         Ok(())
     }
 }