+++ src/main.rs
@@ -5,11 +5,11 @@ use rs_subito_alert::{
     notification::telegram_notifier::TelegramNotifier,
     query_db::query_engine::QueryEngine,
     scraper::{downloader::download_agent::DownloadAgent, scraper_agent::ScraperAgent},
-    serializer::{serializer_agent::SerializerAgent, serializer_api::SerializerApi},
-    telegram_bot::env::TelegramEnvironment,
+    serializer::serializer_agent::SerializerAgent,
+    telegram_bot::{telegram_bot_agent::TelegramBotAgent, telegram_bot_api::TelegramBotApi},
     user_interface::{cli::Cli, user_interface_api::UserInterfaceApi},
 };
-use teloxide::prelude::*;
+use teloxide::{adaptors::DefaultParseMode, prelude::*};
 use tokio::sync::Mutex;
 
 type Application = Subito<
@@ -23,23 +23,11 @@ async fn main() {
     pretty_env_logger::init();
     log::info!("Starting command bot...");
 
-    let (env_serializer, db_serializer) = tokio::join!(
-        SerializerAgent::new(String::from("telegram.json"), None),
-        SerializerAgent::new(String::from("database.json"), None)
-    );
-    let env: TelegramEnvironment = env_serializer
-        .deserialize()
-        .await
-        .ok()
-        .unwrap_or(TelegramEnvironment::default());
-    let bot = Arc::new(Bot::new(env.get_token()));
+    let env_serializer = SerializerAgent::new("telegram.json", None).await;
+    let bot_agent = TelegramBotAgent::new(env_serializer.clone()).await;
+    let bot = bot_agent.get_bot();
     let application = Arc::new(Mutex::new(build_app(Arc::clone(&bot)).await));
-    let cli = Cli::new(
-        Arc::clone(&application),
-        env_serializer,
-        db_serializer,
-        Arc::clone(&bot),
-    );
+    let cli = Cli::new(Arc::clone(&application), bot_agent);
 
     let args: Vec<String> = env::args().collect();
     let skip_cli: bool = match args.get(1) {
@@ -54,9 +42,9 @@ async fn main() {
     }
 }
 
-async fn build_app(bot: Arc<Bot>) -> Application {
-    let env_serializer = Arc::new(SerializerAgent::new(String::from("telegram.json"), None).await);
-    let db_serializer = Arc::new(SerializerAgent::new(String::from("database.json"), None).await);
+async fn build_app(bot: Arc<DefaultParseMode<Bot>>) -> Application {
+    let env_serializer = Arc::new(SerializerAgent::new("telegram.json", None).await);
+    let db_serializer = Arc::new(SerializerAgent::new("database.json", None).await);
     let query_api = Arc::new(Mutex::new(
         QueryEngine::new(Arc::clone(&db_serializer)).await,
     ));