+++ src/notification/telegram_notifier.rs
@@ -1,9 +1,12 @@
 use std::{error::Error, sync::Arc};
 
 use async_trait::async_trait;
-use teloxide::{requests::Requester, Bot};
+use teloxide::{adaptors::DefaultParseMode, requests::Requester, Bot};
 
-use crate::{serializer::serializer_api::SerializerApi, telegram_bot::env::TelegramEnvironment};
+use crate::{
+    scraper::item_result::ItemResult, serializer::serializer_api::SerializerApi,
+    telegram_bot::env::TelegramEnvironment,
+};
 
 use super::notification_api::NotificationApi;
 
@@ -12,14 +15,14 @@ where
     S: SerializerApi<TelegramEnvironment>,
 {
     serializer: Arc<S>,
-    telegram_bot: Arc<Bot>,
+    telegram_bot: Arc<DefaultParseMode<Bot>>,
 }
 
 impl<S> TelegramNotifier<S>
 where
     S: SerializerApi<TelegramEnvironment>,
 {
-    pub fn new(serializer: Arc<S>, bot: Arc<Bot>) -> Self {
+    pub fn new(serializer: Arc<S>, bot: Arc<DefaultParseMode<Bot>>) -> Self {
         Self {
             serializer,
             telegram_bot: bot,
@@ -32,7 +35,7 @@ impl<S> NotificationApi for TelegramNotifier<S>
 where
     S: SerializerApi<TelegramEnvironment> + Send + Sync,
 {
-    async fn notify(&self, item: String) -> Result<(), Box<dyn Error>> {
+    async fn notify(&self, item: &ItemResult) -> Result<(), Box<dyn Error>> {
         let chat_ids = self
             .serializer
             .deserialize()
@@ -41,15 +44,19 @@ where
             .unwrap()
             .get_chat_ids();
         for x in chat_ids {
-            let _ = self.telegram_bot.send_message(x, &item).await?;
+            let _ = self.telegram_bot.send_message(x, item.to_string()).await?;
         }
         Ok(())
     }
 
-    async fn add_user(&self, id: String) -> Result<(), Box<dyn Error>> {
+    async fn add_user(&self, id: &str) -> Result<(), Box<dyn Error>> {
         let mut env = self.serializer.deserialize().await?;
         env.add_user(id)?;
         self.serializer.serialize(&env).await?;
         Ok(())
+    }
+
+    async fn reset(&self) -> Result<(), Box<dyn Error>> {
+        self.serializer.clear().await
     }
 }