+++ tests/test_doubles/application.rs
@@ -42,7 +42,11 @@ impl ApplicationApi for ApplicationDouble {
         todo!()
     }
 
-    async fn add_user(&self, _id: String) -> Result<(), Box<dyn Error>> {
+    async fn add_user(&self, _id: &str) -> Result<(), Box<dyn Error>> {
+        todo!()
+    }
+
+    async fn reset(&self) -> Result<(), Box<dyn Error>> {
         todo!()
     }
 }